Rearange class files and changes on creating NotebookSteps
